+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   `7` lines pre/post
 - [core]: ref: Change way how transports are initialized
 - [core]: ref: Rename `RequestBuffer` to `PromiseBuffer`, also introduce limit
+- [core]: ref: Make sure that captureMessage input is a primitive
 - [browser] fix: Prevent empty exception values
 
 ## 4.4.2